Add files via upload
We aim here at studying the QNM spectra of a spherically symmetric solution of Einstein’s equations, for a dynamical case of Double Pöschl-Teller potential. We first employ the hyperboloidal approach to the problem of QNMs, that incorporates the notion of outgoing boundary conditions in terms of Carter-Penrose diagram, namely black hole horizon and the future null infinity. Then we numerically calculate the QNMs of our system, using Chebyshev spectral methods.